% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINDER-package.R
\docType{package}
\name{BINDER-package}
\alias{BINDER-package}
\alias{BINDER}
\title{The 'BINDER' package.}
\description{
BINDER (BayesIan gene regulatory Networks inferreD via gene coExpression and compaRative genomics) is a hybrid approach, integrating coexpression and comparative genomics to infer prokaryotic regulons.
}
