% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINDER.R
\name{binder}
\alias{binder}
\title{BINDER wrapper function.}
\usage{
binder(proxy_regulon, expression, hyperparams = NULL, inits = NULL,
  O = NULL, delta_CM = "auto", delta_CP = "auto",
  is_coexpression = FALSE, model = "BINDER", n_chains = 1,
  n_draws = 1000, burn_in = 0, thin = 0, n_cores = 1, seed = 1)
}
\arguments{
\item{proxy_regulon}{Data frame comprising columns: \code{regulator}, \code{target_candidate}, \code{ortholog_module_status} (1 if orthologous, 0 otherwise), \code{ME} (1 if high affinity for regulator motif, 0 otherwise) and \code{PE} (1 if orthologous interaction, 0 otherwise).}

\item{expression}{Numerical expression matrix; can be expression matrix with genes represented on rows and experimental conditions represented on columns or symmetric gene-gene coexpression matrix.}

\item{hyperparams}{A named list of user-defined hyperparameters (see \code{?BINDER::run_sampler_binder}, \code{?BINDER::run_sampler_non_auxiliary} and \code{?BINDER::run_sampler_deterministic}).}

\item{inits}{A named list of user-defined initial values to begin each MCMC chain (see \code{?BINDER::run_sampler_binder}, \code{?BINDER::run_sampler_non_auxiliary} and \code{?BINDER::run_sampler_deterministic}).}

\item{O}{Named list where each name is a target candidate and each element is a character vector comprising names of genes that should not be included in the computation of \code{CM} and \code{CP} for the corresponding target candidate name; by default only autocoexpression is excluded.}

\item{delta_CM}{Numerical cutoff point for coexpression module for CM: any values above \code{threshold} are considered to form a coexpression module with \code{target_candidate}; defaults to "auto" which represents the 95th percentile of coexpression scores involving \code{target_candidate}.}

\item{delta_CP}{Numerical cutoff point for coexpression module for CM: any values above \code{threshold} are considered to form a coexpression module with \code{target_candidate}; defaults to "auto" which represents the 95th percentile of coexpression scores involving \code{target_candidate}.}

\item{is_coexpression}{logical; if TRUE, expression matrix is treated as symmetric coexpression matrix.}

\item{model}{Indicates which model should be fit to the data: one of "BINDER", "non_auxiliary" or "deterministic".}

\item{n_chains}{Number of MCMC chains to run; defaults to 1.}

\item{n_draws}{Number of MCMC draws to run; defaults to 1000.}

\item{burn_in}{Number of MCMC draws to discard from the beginning of each chain; defaults to 0.}

\item{thin}{Number of MCMC draws to discard between each accepted MCMC draw for each chain; defaults to 0.}

\item{seed}{Seed for random number generation; defaults to 1.}
}
\value{
A list containing the MCMC draws for each model parameter for each chain for each putative regulor in \code{proxy_regulon}.
}
\description{
BINDER wrapper function.
}
