% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{compute_CM}
\alias{compute_CM}
\title{Compute CM value for target candidate.}
\usage{
compute_CM(target_candidate, target_candidates, ME_module, coexpression, o,
  threshold = "auto")
}
\arguments{
\item{target_candidate}{Target candidate for which CM value is computed.}

\item{target_candidates}{Character vector containing all target candidates for regulator.}

\item{ME_module}{Character vector comprising all target candidates with a high affinity for regulator motif.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{o}{Character vector comprising names of genes which should not be included in the computation of `CM` for `target_candidate`.}

\item{threshold}{Numerical cutoff point: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
CM value for `target_candidate`.
}
\description{
Compute CM value for target candidate.
}
