% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{compute_coexpression}
\alias{compute_coexpression}
\title{Derive symmetric coexpression matrix from expression matrix with genes represented on rows and experimental conditions represented on columns.}
\usage{
compute_coexpression(expression, target_candidates)
}
\arguments{
\item{expression}{Numerical expression matrix with genes represented on rows and experimental conditions represented on columns.}

\item{target_candidates}{Character vector containing all target candidates for regulator.}
}
\value{
Derived numerical symmetric coexpression matrix.
}
\description{
Derive symmetric coexpression matrix from expression matrix with genes represented on rows and experimental conditions represented on columns.
}
