% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{compute_CP}
\alias{compute_CP}
\title{Compute CP value for target candidate.}
\usage{
compute_CP(target_candidate, target_candidates, PE_module, ortholog_module,
  coexpression, o = NULL, threshold = "auto")
}
\arguments{
\item{target_candidate}{Target candidate for which CP value is computed.}

\item{target_candidates}{Character vector containing all target candidates for regulator.}

\item{PE_module}{Character vector comprising all target candidates with precedent orthologous interaction with regulator.}

\item{ortholog_module}{Character vector comprising all orthologous target candidates.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{o}{Character vector comprising names of genes which should not be included in the computation of \code{CP} for \code{target_candidate}.}

\item{threshold}{Numerical cutoff point: any values above \code{threshold} are considered to form a coexpression module with \code{target_candidate}; defaults to "auto" which represents the 95th percentile of coexpression scores involving \code{target_candidate}.}
}
\value{
CP value for \code{target_candidate}.
}
\description{
Compute CP value for target candidate.
}
