% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{compute_CPs}
\alias{compute_CPs}
\title{Compute CP values for all target candidates.}
\usage{
compute_CPs(target_candidates, PE_module, ortholog_module, coexpression, O,
  threshold = "auto")
}
\arguments{
\item{target_candidates}{Character vector containing all target candidates for regulator.}

\item{PE_module}{Character vector comprising all target candidates with precedent orthologous interaction with regulator.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{O}{Named list where each name is a target candidate and each element is a character vector comprising names of genes that should not be included in the computation of `CP` for the corresponding target candidate name; by default only autocoexpression is excluded.}

\item{threshold}{Numerical cutoff point: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
Numerical vector containing CP values for `target_candidates`.
}
\description{
Compute CP values for all target candidates.
}
