% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{find_coexpression_module}
\alias{find_coexpression_module}
\title{Generate coexpression module for target candidate.}
\usage{
find_coexpression_module(target_candidate, coexpression, operon,
  threshold = "auto")
}
\arguments{
\item{target_candidate}{Target candidate for which coexpression module is generated.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{operon}{Character vector comprising genes of operon to which `target_candidate` belongs.}

\item{threshold}{Numerical cutoff point: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
Character vector containing genes sharing a coexpression with `target_candidate` greater than `threshold`.
}
\description{
Generate coexpression module for target candidate.
}
