% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for use in BINDER model.}
\usage{
prepare_data(proxy_structure)
}
\arguments{
\item{proxy_structure}{Data frame returned by `build_proxy_structure` implementing augmented structure comprising columns: `regulator`, `target_candidate`, `ortholog_module_status` (1 if orthologous, 0 otherwise), `ME` (1 if high affinity for regulator motif, 0 otherwise), `PE` (1 if orthologous interaction, 0 otherwise), `CM` and `CP`.}
}
\value{
List comprising `regulator`, `target_candidate`, `N` (number of regulator-target observations), `K` (number of variables in the auxiliary stratum := 2 (i.e. ME and PE)), `M` (number of variables in the primary stratum := 2 (i.e. CM and CP)), `X` (matrix comprising the variables in the auxiliary stratum (ME, PE)), `Y` (matrix comprising the variables in the primary stratum (CM, CP)).
}
\description{
Prepare data for use in BINDER model.
}
