% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINDER.R
\name{run_binder}
\alias{run_binder}
\title{Run BINDER model using prepared data.}
\usage{
run_binder(prepared_data, mu_zeta = 0, sigma_zeta = 3, mu_tau = c(0,
  0), sigma_tau = c(3, 3), mu_phi = 0, sigma_phi = 3, mu_psi = c(0,
  0), sigma_psi = c(3, 3), model_summary = FALSE,
  model_summary_parameters, model_name = "anon_model", fit = NA,
  chains = 4, iter = 2000, warmup = floor(iter/2), thin = 1,
  init = "random", seed = sample.int(.Machine$integer.max, 1),
  algorithm = c("NUTS", "HMC", "Fixed_param"), control = NULL,
  sample_file = NULL, diagnostic_file = NULL, save_dso = TRUE,
  verbose = FALSE, include = TRUE, cores = getOption("mc.cores", 1L),
  open_progress = interactive() && !isatty(stdout()) &&
  !identical(Sys.getenv("RSTUDIO"), "1"), chain_id, init_r,
  test_grad = FALSE, append_samples, refresh = max(iter/10, 1),
  enable_random_init, save_warmup = TRUE, boost_lib = NULL,
  eigen_lib = NULL)
}
\arguments{
\item{prepared_data}{A list generated by `prepare_data` function.}

\item{mu_zeta}{Prior mean for zeta; the default is 0.}

\item{sigma_zeta}{Prior standard deviation for zeta; the default is 3.}

\item{mu_tau}{Prior mean vector for tau; the default is c(0,0).}

\item{sigma_tau}{Prior standard deviation vector for tau; the default is c(3,3).}

\item{mu_phi}{Prior mean for phi; the default is 0.}

\item{sigma_phi}{Prior standard deviation for phi; the default is 3.}

\item{mu_psi}{Prior mean vector for psi; the default is c(0,0).}

\item{sigma_psi}{Prior standard deviation vector for psi; the default is c(3,3).}

\item{model_summary}{logical; if TRUE, a model summary of parameter estimates is printed upon completion of BINDER run; defaults to FALSE.}

\item{model_summary_parameters}{A vector of parameters to include in model summary.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
A list with the following components:
  \itemize{
    \item regulator - the name of the regulator provided to BINDER.
    \item target_candidate - a vector providing the names of each target candidate provided to BINDER.
    \item ME - a vector providing ME values for each target candidate provided to BINDER.
    \item PE - a vector providing PE values for each target candidate provided to BINDER.
    \item CM - a vector providing CM values for each target candidate provided to BINDER.
    \item CP - a vector providing CP values for each target candidate provided to BINDER.
    \item mean_theta - a vector providing posterior mean values for theta for each target candidate provided to BINDER.
    \item theta_interval - a matrix providing posterior 0.025, 0.5 and 0.975 quantiles for theta for each target candidate provided to BINDER.
    \item mean_gamma - a vector providing posterior mean values for gamma for each target candidate provided to BINDER.
    \item gamma_interval - a matrix providing posterior 0.025 and 0.975 quantiles for gamma for each target candidate provided to BINDER.
    \item mean_raw_gamma - a vector providing posterior mean values for raw_gamma for each target candidate provided to BINDER.
    \item raw_gamma_interval - a matrix providing posterior 0.025 and 0.975 quantiles for raw_gamma for each target candidate provided to BINDER.
    \item mean_phi - posterior mean for phi as estimated by BINDER.
    \item phi_interval - a vector providing posterior 0.025 and 0.975 quantiles for phi as estimated by BINDER.
    \item mean_zeta - posterior mean for zeta as estimated by BINDER.
    \item zeta_interval - a vector posterior 0.025 and 0.975 quantiles for zeta as estimated by BINDER.
    \item mean_tau - a vector providing posterior mean values for tau as estimated by BINDER.
    \item tau_interval - a matrix providing posterior 0.025 and 0.975 quantiles for tau as estimated by BINDER.
    \item mean_psi - a vector providing posterior mean values for psi as estimated by BINDER.
    \item psi_interval - a matrix providing posterior 0.025 and 0.975 quantiles for psi as estimated by BINDER.
    \item posterior_samples_theta - a matrix providing posterior samples for each target candidate provided to BINDER.
    \item posterior_samples_psi - a matrix providing posterior samples for each psi element as estimated by BINDER.
    \item model_object - object of class `stanfit` returned by `rstan::sampling`.
  }
}
\description{
Run BINDER model using prepared data.
}
