% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINDER.R
\name{run_sampler_non_auxiliary}
\alias{run_sampler_non_auxiliary}
\title{Run non-auxiliary BINDER model using prepared data.}
\usage{
run_sampler_non_auxiliary(prepared_data, hyperparams = NULL,
  inits = NULL, n_chains = 1, n_draws = 1000, burn_in = 0,
  thin = 0, seed = 1)
}
\arguments{
\item{prepared_data}{A named list generated by \code{prepare_data} function comprising model input data.}

\item{hyperparams}{A named list of user-defined hyperparameters; this list should contain some or all of the following named values: alpha_psi_CM (prior shape parameter for psi_CM (inverse gamma distribution)), beta_psi_CM (prior scale parameter for psi_CM (inverse gamma distribution)), alpha_psi_CP (prior shape parameter for psi_CP (inverse gamma distribution)), beta_psi_CP (prior scale parameter for psi_CP (inverse gamma distribution))); for a given hyperparameter, default values are used if no user-defined value is provided to the function call (alpha_psi_CM := 1.5, beta_psi_CM := 1.5, alpha_psi_CP := 1.5, beta_psi_CP := 1.5).}

\item{inits}{A named list of user-defined initial values to begin each MCMC chain (each element of this list should comprise a vector of length equal to the number of chains and each element of the vector should correspond to the initial value for that parameter for each chain); this list should contain some or all of the following named values: theta (defaults to randomly generated value from uniform(0.0001, 0.9999) distribution for each target candidate's corresponding theta), psi_CM (defaults to randomly generated value from prior inverse-gamma distribution for psi_CM), psi_CP (defaults to randomly generated value from prior inverse-gamma distribution for psi_CP).}

\item{n_chains}{Number of MCMC chains to run; defaults to 1.}

\item{n_draws}{Number of MCMC draws to run; defaults to 1000.}

\item{burn_in}{Number of MCMC draws to discard from the beginning of each chain; defaults to 0.}

\item{thin}{Number of MCMC draws to discard between each accepted MCMC draw for each chain; defaults to 0.}

\item{seed}{Seed for random number generation; defaults to 1.}
}
\value{
A list containing the MCMC draws for each model parameter for each chain.
}
\description{
Run non-auxiliary BINDER model using prepared data.
}
