% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINDER.R
\name{summary.posteriorDraws}
\alias{summary.posteriorDraws}
\title{\code{summary} function for class \code{posteriorDraws}.}
\usage{
\method{summary}{posteriorDraws}(x, target_candidates = c("all"))
}
\arguments{
\item{x}{Object of class \code{posteriorDraws} such as that returned by \code{BINDER::binder} function call.}

\item{target_candidates}{Named list where each element comprises a vector subset of target candidate IDs of interest and each list element is named by the regulator ID to which the target candidate subset pertains; if \code{target_candidates} is not specified, default behaviour returns summaries for all regulator target candidate pairs considered in the model run.}
}
\value{
Summaries (mean, std.dev, 0%, 25%, 50%, 75% and 100% quantiles) for the posterior distribution of theta for each regulator and target candidate pair of interest.
}
\description{
\code{summary} function for class \code{posteriorDraws}.
}
