% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINDER-package.R
\docType{package}
\name{BINDER-package}
\alias{BINDER-package}
\alias{BINDER}
\title{The 'BINDER' package.}
\description{
BINDER (BayesIan gene regulatory Networks inferreD via gene coExpression and compaRative genomics) is a hybrid approach, binding coexpression and comparative genomics to infer prokaryotic regulons.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
