% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{build_proxy_structure}
\alias{build_proxy_structure}
\title{Build augmented structure comprising CM and CP values from coexpression, ME and PE values.}
\usage{
build_proxy_structure(ortholog_ME_PE_structure, coexpression, operons,
  delta_CM = "auto", delta_CP = "auto")
}
\arguments{
\item{ortholog_ME_PE_structure}{Data frame comprising columns: `regulator`, `target_candidate`, `ortholog_module_status` (1 if orthologous, 0 otherwise), `ME` (1 if high affinity for regulator motif, 0 otherwise) and `PE` (1 if orthologous interaction, 0 otherwise).}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{operons}{Data frame in DOOR format comprising operon annotations for all target candidates in `ortholog_ME_PE_structure`.}

\item{delta_CM}{Numerical cutoff point for coexpression module for CM: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}

\item{delta_CP}{Numerical cutoff point for coexpression module for CM: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
Data frame implementing augmented structure comprising columns: `regulator`, `target_candidate`, `ortholog_module_status` (1 if orthologous, 0 otherwise), `ME` (1 if high affinity for regulator motif, 0 otherwise), `PE` (1 if orthologous interaction, 0 otherwise), `CM` and `CP`.
}
\description{
Build augmented structure comprising CM and CP values from coexpression, ME and PE values.
}
