% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{build_proxy_structure}
\alias{build_proxy_structure}
\title{Build augmented structure comprising CM and CP values from coexpression, ME and PE values.}
\usage{
build_proxy_structure(proxy_regulon, coexpression, O, delta_CM = "auto",
  delta_CP = "auto")
}
\arguments{
\item{proxy_regulon}{Data frame comprising columns: `regulator`, `target_candidate`, `ortholog_module_status` (1 if orthologous, 0 otherwise), `ME` (1 if high affinity for regulator motif, 0 otherwise) and `PE` (1 if orthologous interaction, 0 otherwise).}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{O}{Named list where each name is a target candidate and each element is a character vector comprising names of genes that should not be included in the computation of `CM` and `CP` for the corresponding target candidate name; by default only autocoexpression is excluded.}

\item{delta_CM}{Numerical cutoff point for coexpression module for CM: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}

\item{delta_CP}{Numerical cutoff point for coexpression module for CM: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
Data frame implementing augmented structure comprising columns: `regulator`, `target_candidate`, `ortholog_module_status` (1 if orthologous, 0 otherwise), `ME` (1 if high affinity for regulator motif, 0 otherwise), `PE` (1 if orthologous interaction, 0 otherwise), `CM` and `CP`.
}
\description{
Build augmented structure comprising CM and CP values from coexpression, ME and PE values.
}
