% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{compute_CMs}
\alias{compute_CMs}
\title{Compute CM values for all target candidates.}
\usage{
compute_CMs(target_candidates, ME_module, coexpression, operons,
  threshold = "auto")
}
\arguments{
\item{target_candidates}{Character vector containing all target candidates for regulator.}

\item{ME_module}{Character vector comprising all target candidates with a high affinity for regulator motif.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{operons}{Data frame in DOOR format comprising operon annotations for all genes in `target_candidates`.}

\item{threshold}{Numerical cutoff point: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
Numerical vector containing CM values for `target_candidates`.
}
\description{
Compute CM values for all target candidates.
}
