% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{compute_CP}
\alias{compute_CP}
\title{Compute CP value for target candidate.}
\usage{
compute_CP(target_candidate, target_candidates, PE_module, ortholog_module,
  coexpression, operon, threshold = "auto")
}
\arguments{
\item{target_candidate}{Target candidate for which CP value is computed.}

\item{target_candidates}{Character vector containing all target candidates for regulator.}

\item{PE_module}{Character vector comprising all target candidates with precedent orthologous interaction with regulator.}

\item{ortholog_module}{Character vector comprising all orthologous target candidates.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{operon}{Character vector comprising genes of operon to which `target_candidate` belongs.}

\item{threshold}{Numerical cutoff point: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
CP value for `target_candidate`.
}
\description{
Compute CP value for target candidate.
}
