% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{compute_CPs}
\alias{compute_CPs}
\title{Compute CP values for all target candidates.}
\usage{
compute_CPs(target_candidates, PE_module, ortholog_module, coexpression,
  operons, threshold = "auto")
}
\arguments{
\item{target_candidates}{Character vector containing all target candidates for regulator.}

\item{PE_module}{Character vector comprising all target candidates with precedent orthologous interaction with regulator.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{operons}{Data frame in DOOR format comprising operon annotations for all genes in `target_candidates`.}

\item{threshold}{Numerical cutoff point: any values above `threshold` are considered to form a coexpression module with `target_candidate`; defaults to "auto" which represents the 95th percentile of coexpression scores involving `target_candidate`.}
}
\value{
Numerical vector containing CP values for `target_candidates`.
}
\description{
Compute CP values for all target candidates.
}
