% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression.R
\name{find_coexpression_module}
\alias{find_coexpression_module}
\title{Generate coexpression module for target candidate.}
\usage{
find_coexpression_module(target_candidate, coexpression, o = NULL,
  threshold = "auto")
}
\arguments{
\item{target_candidate}{Target candidate for which coexpression module is generated.}

\item{coexpression}{Numerical symmetric matrix providing gene coexpression values.}

\item{o}{Character vector comprising names of genes which should not be included in the computation of \code{CM} and \code{CP} for \code{target_candidate}.}

\item{threshold}{Numerical cutoff point: any values above \code{threshold} are considered to form a coexpression module with \code{target_candidate}; defaults to "auto" which represents the 95th percentile of coexpression scores involving \code{target_candidate}.}
}
\value{
Character vector containing genes sharing a coexpression with \code{target_candidate} greater than \code{threshold}.
}
\description{
Generate coexpression module for target candidate.
}
